+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # FastCGI Cache Bust Changelog
 
+## 1.0.3 - 2017.11.27
+### Changed
+* Don't bust the FastCGI Cache unless the element being saved is ENABLED or LIVE
+* Don't bust the FastCGI Cache for certain custom ElementTypes
+
 ## 1.0.2 - 2017.08.12
 ### Changed
 * Added 'FastCGI Cache' to the list of things that can be cleared via Craft's Clear Caches tool